wpilib: Rename opmoderobot module to opmodes
